requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
autopep8==1.5.7
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.0
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
Django==5.1.4
django-cors-headers==3.8.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.0
numpy==2.2.1
Pillow==10.4
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.7.0
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.3.0
python-dateutil==2.9.0.post0
pytz==2021.1
redis==5.2.1
requests==2.32.3
service-identity==24.2.0
six==1.17.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.2.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
zope.interface==7.2
